+++ client/src/router.tsx
@@ -1,9 +1,8 @@
 import { createBrowserRouter } from "react-router-dom";
-import App from "./App.tsx";
 
 export const router = createBrowserRouter([
     {
         path: "/",
-        element: <App />,
+        element: <></>,
     },
 ]);
